refactor(trusty-mpm): move SessionManager setters to setters.rs
manager.rs was at the 500-SLOC production cap; move the post-creation
field setters (mark_errored, set_workspace, set_pending_decision,
set_source_id, set_deliverable_id, clear_pending_decision) to a sibling
setters.rs, following the stop.rs/create.rs precedent, to make room for
#7087 slice 1b's residency-generation tracking.

Refs #7087
